requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.3
bitarray==3.0.0
black==24.10.0
bleach==6.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
cytoolz==1.0.0
decorator==5.0.5
defusedxml==0.7.1
docopt==0.6.2
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
executing==2.1.0
fastjsonschema==2.21.1
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
ipython==8.12.3
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.0.0
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyunormalize==16.0.0
pytest==8.4.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
ruff==0.4.4
rlp==4.0.1
rpds-py==0.22.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlalchemy_utils==0.41.2
stack-data==0.6.3
tinycss2==1.4.0
toolz==1.0.0
tornado==6.5.1
traitlets==5.14.3
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
web3==7.6.0
webencodings==0.5.1
websockets==13.1
yarg==0.1.9
yarl==1.18.3
